+++ pubspec.yaml
@@ -54,6 +54,7 @@ dependencies:
   firebase_storage: ^11.2.6
   image_picker: ^1.0.4
   path_provider: ^2.1.1
+  uuid: ^4.0.0
 
 dev_dependencies:
   flutter_test:
